fix cram review
